+++ R/plotExplanatoryHeatmap.R
@@ -8,7 +8,8 @@ heatmapClasses <- function(pl,
                            distanceMeasure, 
                            clusterMethod, 
                            featureNames,
-                           dendrogram){
+                           dendrogram,
+                           featureLimit){
   pl %>%
     map(~{
       r <- .
@@ -22,6 +23,10 @@ heatmapClasses <- function(pl,
       
       feat <- r$Feature %>%
         unique()
+      
+      if (length(feat) > featureLimit){
+        feat <- feat[1:featureLimit]
+      }
       
       d <- x %>%
         keepClasses(cls = pred,
@@ -101,6 +106,8 @@ heatmapClasses <- function(pl,
       }
       
       if (isTRUE(dendrogram)) {
+        offset <- 1 / length(feat) * 0.5
+        
         dend_plot <- ggplot() +
           geom_segment(
             data = dend$segments,
@@ -108,7 +115,7 @@ heatmapClasses <- function(pl,
           scale_x_reverse() +
           scale_y_continuous(breaks = seq_along(dend$labels$label), 
                              labels = dend$labels$label,position = 'right',
-                             expand = c(0.005,0)) +
+                             expand = c(offset,offset)) +
           theme_minimal(base_size = 14) +
           theme(axis.text.x = element_blank(),
                 panel.grid = element_blank(),
@@ -133,7 +140,8 @@ heatmapRegression <- function(pl,
                               distanceMeasure, 
                               clusterMethod, 
                               featureNames, 
-                              dendrogram){
+                              dendrogram,
+                              featureLimit){
   pl %>%
     map(~{
       
@@ -141,6 +149,10 @@ heatmapRegression <- function(pl,
       
       feat <- .x$Feature %>%
         unique()
+      
+      if (length(feat) > featureLimit){
+        feat <- feat[1:featureLimit]
+      }
       
       p <- sym(response)
       
@@ -209,6 +221,8 @@ heatmapRegression <- function(pl,
       }
       
       if (isTRUE(dendrogram)) {
+        offset <- 1 / length(feat) * 0.5
+        
         dend_plot <- ggplot() +
           geom_segment(
             data = dend$segments,
@@ -216,7 +230,7 @@ heatmapRegression <- function(pl,
           scale_x_reverse() +
           scale_y_continuous(breaks = seq_along(dend$labels$label), 
                              labels = dend$labels$label,position = 'right',
-                             expand = c(0.005,0)) +
+                             expand = c(offset,offset)) +
           theme_minimal(base_size = 14) +
           theme(axis.text.x = element_blank(),
                 panel.grid = element_blank(),
@@ -244,7 +258,12 @@ heatmapRegression <- function(pl,
 #' @param clusterMethod clustering method to use. See details
 #' @param featureNames should feature names be plotted?
 #' @param dendrogram TRUE/FALSE. Should the dendrogram be plotted?
+#' @param featureLimit The maximum number of features to plot
 #' @param ... arguments to pass to the appropriate method
+#' @details 
+#' Distance measures can be one of any that can be used for the `method` argument of [dist()].
+#'
+#' Cluster methods can be one of any that can be used for the `method` argument of [hclust()].
 #' @examples
 #' library(metaboData)
 #' x <- analysisData(data = abr1$neg[,200:300],info = abr1$fact)
@@ -282,10 +301,16 @@ setMethod('plotExplanatoryHeatmap',
                    distanceMeasure = "euclidean", 
                    clusterMethod = 'ward.D2', 
                    featureNames = TRUE, 
-                   dendrogram = TRUE){
+                   dendrogram = TRUE,
+                   featureLimit = Inf){
             
             res <- x %>%
-              explanatoryFeatures()
+              explanatoryFeatures(threshold = threshold)
+            
+            if (nrow(res) < 1){
+              message('No explanatory features found at this threshold.')
+              return()
+            }
             
             pl <- res %>%
               base::split(.$Response)
@@ -299,7 +324,8 @@ setMethod('plotExplanatoryHeatmap',
                 distanceMeasure = distanceMeasure, 
                 clusterMethod = clusterMethod, 
                 featureNames = featureNames,
-                dendrogram = dendrogram)
+                dendrogram = dendrogram,
+                featureLimit = featureLimit)
             }
             
             if (x@type == 'linear regression') {
@@ -311,7 +337,8 @@ setMethod('plotExplanatoryHeatmap',
                 distanceMeasure = distanceMeasure, 
                 clusterMethod = clusterMethod, 
                 featureNames = featureNames,
-                dendrogram = dendrogram)
+                dendrogram = dendrogram,
+                featureLimit = featureLimit)
             }
             
             feat <- res$Feature %>% 
@@ -321,10 +348,10 @@ setMethod('plotExplanatoryHeatmap',
               'Explanatory features had a P value below a threshold of ',
               threshold,'.')
             
-            if (length(feat) > 500) {
+            if (length(feat) > featureLimit) {
               caption <- str_c(
                 caption,'\n',
-                'Number of features capped at top 500.')
+                str_c('Number of features capped at top ',featureLimit,'.'))
             }
             
             pl <- wrap_plots(pl) + 
@@ -347,7 +374,8 @@ setMethod('plotExplanatoryHeatmap',
                    distanceMeasure = "euclidean",
                    clusterMethod = 'ward.D2',
                    featureNames = TRUE, 
-                   dendrogram = TRUE){
+                   dendrogram = TRUE,
+                   featureLimit = Inf){
             
             if (x@type == 'unsupervised') {
               stop('Cannot plot heatmap for unsupervised random forest.')
@@ -357,19 +385,25 @@ setMethod('plotExplanatoryHeatmap',
                                           metric = metric,
                                           threshold = threshold)
             
+            if (nrow(explan) < 1){
+              message('No explanatory features found at this threshold.')
+              return()
+            }
+            
             pl <- explan %>%
               base::split(.$Response)
             
             if (x@type == 'classification') {
               pl <- heatmapClasses(
                 pl,
-                x, 
+                x,
                 threshold = threshold, 
                 title = title,
                 distanceMeasure = distanceMeasure, 
                 clusterMethod = clusterMethod, 
                 featureNames = featureNames,
-                dendrogram = dendrogram)
+                dendrogram = dendrogram,
+                featureLimit = featureLimit)
             }
             
             if (x@type == 'regression') {
@@ -381,7 +415,8 @@ setMethod('plotExplanatoryHeatmap',
                 distanceMeasure = distanceMeasure, 
                 clusterMethod = clusterMethod, 
                 featureNames = featureNames,
-                dendrogram = dendrogram)
+                dendrogram = dendrogram,
+                featureLimit = featureLimit)
             }
             
             feat <- explan$Feature %>% 
@@ -397,10 +432,10 @@ setMethod('plotExplanatoryHeatmap',
               'Explanatory features had an importance value ',direction, ' a threshold of ',
               threshold,'.')
             
-            if (length(feat) > 500) {
+            if (length(feat) > featureLimit) {
               caption <- str_c(
                 caption,'\n',
-                'Number of features capped at top 500.')
+                str_c('Number of features capped at top ',featureLimit,'.'))
             }
             
             pl <- wrap_plots(pl) + 
@@ -420,7 +455,8 @@ setMethod('plotExplanatoryHeatmap',
                    threshold = 0.05, 
                    distanceMeasure = "euclidean",
                    clusterMethod = 'ward.D2',
-                   featureNames = TRUE){
+                   featureNames = TRUE,
+                   featureLimit = Inf){
             object_classes <- x %>%
               map_chr(class)
             
@@ -440,7 +476,8 @@ setMethod('plotExplanatoryHeatmap',
                 title = .x,
                 distanceMeasure = distanceMeasure, 
                 clusterMethod = clusterMethod,
-                featureNames = featureNames
+                featureNames = featureNames,
+                featureLimit = featureLimit
               )
               })
           }
@@ -454,12 +491,14 @@ setMethod('plotExplanatoryHeatmap',
                    threshold = 0.05, 
                    distanceMeasure = "euclidean", 
                    clusterMethod = 'ward.D2', 
-                   featureNames = TRUE){
+                   featureNames = TRUE,
+                   featureLimit = Inf){
             x %>%
               analysisResults(element = 'modelling') %>%
               plotExplanatoryHeatmap(threshold = threshold, 
                                      distanceMeasure = distanceMeasure, 
                                      clusterMethod = clusterMethod, 
-                                     featureNames = featureNames) 
+                                     featureNames = featureNames,
+                                     featureLimit = featureLimit) 
           }
 )