+++ tests/testthat/test-plotRSD.R
@@ -3,29 +3,27 @@ library(metaboData)
 context('plotRSD')
 
 test_that('plotRSD works for Analysis class',{
-  d <- new('Analysis')
-  
-  raw(d) <- analysisData(abr1$neg[,300:400],abr1$fact)
-  
-  p <- analysisParameters('pre-treatment')
-  
-  parameters(p,'pre-treatment') <- list(
-    keep = list(classes = list(cls = 'day',classes = '5')),
-    occupancyFilter = list(maximum = list(cls = 'day',occupancy = 2/3)),
-    impute = list(all = list(occupancy = 2/3,
-                             parallel = 'variables',
-                             seed = 1234))
-  )
-  
-  pl <- plotRSD(d,cls = 'day',QCidx = '5',QCparameters = p)
-  
-  expect_identical(class(pl),c('patchwork','gg','ggplot'))
-})
-
-test_that('plotRSD works for Analysis class',{
   d <- analysisData(abr1$neg,abr1$fact)
   
   pl <- plotRSD(d,cls = 'day')
   
-  expect_identical(class(pl),c('patchwork','gg','ggplot'))
+  expect_s3_class(pl,'patchwork')
+})
+
+test_that('plotRSD errors correctly when numeric sample info column specified',{
+  d <- analysisData(abr1$neg,abr1$fact)
+  
+  pl <- plotRSD(d,cls = 'day')
+  
+  expect_error(plotRSD(d,cls = 'class'))
+})
+
+test_that('plotRSD works for Analysis class',{
+  d <- new('Analysis')
+  
+  raw(d) <- analysisData(abr1$neg[,300:400],abr1$fact)
+  
+  pl <- plotRSD(d,cls = 'day')
+  
+  expect_s3_class(pl,'patchwork')
 })